
\name{diverse}
\docType{data}
\alias{diverse}
\title{(Dataset of requirements for Colombia)-diverse}

\description{
Dataset containing the quantity of foods to be selected per groups according to GABAS.

----

Conjunto de datos que contiene la cantidad de alimentos a seleccionar por grupos o subgrupos según GABAS.
}

\usage{
diverse
}

\format{
Dataset with two columns ("Group" and "Number") and ten rows (food groups).

----

Conjunto de datos con dos columnas ("Group" y "Number") y diez filas (grupos de alimentos).
}

\examples{

library(Foodprice)
head(diverse)

}

\source{
- GABAS: Instituto Colombiano de Bienestar Familiar. (2020). Guías Alimentarias Basadas en Alimentos para la población colombiana mayor de 2 años. En ICBF/FAO (2da Edic.). Recovered from https://www.minsalud.gov.co/sites/rid/Lists/BibliotecaDigital/RIDE/VS/PP/SNA/guiasalimentariasbasadas-en-alimentos.pdf
}


