\name{diverse_example}
\docType{data}
\alias{diverse_example}
\title{(Example dataset)-diverse_example}

\description{

The dataset contains the quantity of foods to be selected. The food groups have arbitrary names as the database was created from 'diverse' for purely illustrative purposes.

----

El conjunto de datos incluye la cantidad de alimentos a seleccionar. Los grupos de alimentos tienen nombres arbitrarios, ya que la base de datos se creó a partir de 'diverse' con un propósito puramente ilustrativo.

}

\format{
Dataset with two columns ("Group" and "Number") and ten rows (food groups).

----

Conjunto de datos con dos columnas ("Group" y "Number") y diez filas (grupos de alimentos).
}

\usage{
diverse_example
}


\examples{

library(Foodprice)
head(diverse_example)
}