
\name{EER_example}
\docType{data}
\alias{EER_example}
\title{(Example dataset)-EER_example}

\description{

Sample dataset containing energy requirements for arbitrary demographic groups. These values match those established for the entire Colombia in the 'EER' data.

----    

Conjunto de datos de ejemplo que contiene los requisitos energéticos para grupos demográficos arbitrarios. Estos valores coinciden con los establecidos para toda Colombia en los datos de 'EER'

}

\format{

Dataset with 22 demographic groups (rows) and 3 variables (columns), namely, "Sex" (0 if male and 1 if female), "Age" (Demographic group), and "Energy" (energy requirement for each gender and demographic group).

----

Conjunto de datos con 22 grupos demográficos (filas) y 3 variables (columnas), a saber, "Sex" (0 si es masculino y 1 si es femenino), "Age" (grupo demográfico) y "Energy" (requerimiento energético para cada sexo y grupo demográfico).
}

\usage{
EER_example
}

\examples{

library(Foodprice)
head(EER_example)

}

