
\name{EER_LL_example}
\docType{data}
\alias{EER_LL_example}
\title{(Example dataset)-EER_LL_example}

\description{
Sample dataset with energy requirements and minimum requirements for various nutrients in arbitrary demographic groups and genders. The requirements are the same as those established for all of Colombia in the 'EER_LL' data.

----

Conjunto de datos de ejemplo con los requerimientos energéticos y requerimientos mínimos para varios nutrientes en grupos demográficos y sexos arbitrarios. Los requisitos son los mismos que los establecidos para todo Colombia en los datos 'EER_LL'.

}

\format{

Dataset with 22 demographic groups (rows) and 19 variables (columns). The columns "Sex" (Sex) and "Age" (Age) correspond to the genders and demographic groups, respectively; the remaining columns represent the minimum requirements associated with various nutrients.

----

Conjunto de datos con 22 grupos demográficos (filas) y 19 variables (columnas). Las columnas "Sex" (Sexo) y "Age" (Edad) corresponden a los géneros y grupos demográficos respectivamente; las columnas restantes representan los requisitos mínimos asociados con varios nutrientes.

}

\usage{
EER_LL_example
}

\examples{

library(Foodprice)
head(EER_LL_example)

}

