
\name{UL_example}
\alias{UL_example}
\title{(Example dataset)-UL_example}

\description{

Dataset with maximum requirements for various nutrients in different arbitrary demographic groups by gender. The values of the energy requirements are the same as those in the 'UL' dataset.

----

Conjunto de datos con los requerimientos máximos para varios nutrientes en diferentes grupos demográficos arbitrarios por sexo. Los valores de los requisitos energéticos son iguales a los del conjunto de datos 'UL'.
}

\usage{
UL_example
}

\format{

Dataset with 22 demographic groups (rows) and 19 variables (columns). The columns "Sex" (Sex) and "Age" (Age) correspond to the genders and demographic groups, respectively; the remaining columns represent the maximum requirements associated with various nutrients.

----

Conjunto de datos con 22 grupos demográficos (filas) y 19 variables (columnas). Las columnas "Sex" (Sexo) y "Age" (Edad) corresponden a los sexos y grupos demográficos, respectivamente; las columnas restantes representan los requisitos máximos asociados a varios nutrientes.

}

\examples{

library(Foodprice)
head(UL_example)

}

